Moves files around to adhere to adr
