+++ src/app/components/movie-search-bar/index.ts
@@ -1,0 +1,1 @@
+export * from './movie-search-bar'
